requirements.txt
deepgram-sdk>=3.0.0,<4.0.0
elevenlabs>=1.0.0
groq>=0.11.0
python-dotenv>=1.0.0
requests>=2.31.0
websockets>=12.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
PyMuPDF>=1.24.0
